Move contents type definition
