+++ packages/zenn-cli/src/client/components/chapters/show/ChapterHeader.tsx
@@ -1,8 +1,7 @@
 import React, { useMemo } from 'react';
 import { MdOutlineArrowBackIos } from 'react-icons/md';
 import styled from 'styled-components';
-import { Book, Chapter } from '../../../../common/types';
-import { validateBookChapter } from 'zenn-validator';
+import { Book, Chapter, validateBookChapter } from 'zenn-validator';
 import { ContentContainer } from '../../ContentContainer';
 import { PropertyRow } from '../../PropertyRow';
 import { LinkBook } from '../../Routes';
