+++ packages/zenn-cli/src/server/lib/books.ts
@@ -22,6 +22,8 @@ import {
   BookMeta,
   Chapter,
   ChapterMeta,
+} from 'zenn-validator';
+import {
   ItemSortType,
 } from '../../common/types';
